+++ android/app/src/main/java/id/my/rizalanggoro/arta/feature/home/presentation/gold/HomeGoldScreen.kt
@@ -105,7 +105,7 @@ private fun Preview() {
                             grams = 3.3.toBigDecimal(),
                             id = 1,
                             notes = "",
-                            price = 1500000.toBigDecimal(),
+                            price = 1500000,
                             type = "jewelry",
                             updatedAt = "2026-05-25T14:38:00.000+07:00",
                             walletId = 1
